requirements/requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

annotated-types==0.6.0
colorama==0.4.6; sys_platform == "win32" or platform_system == "Windows"
gitdb==4.0.11
gitpython==3.1.42
loguru==0.7.2
pendulum==3.0.0
pydantic==2.6.4
pydantic-core==2.16.3
python-dateutil==2.9.0.post0
six==1.16.0
smmap==5.0.1
time-machine==2.14.0; implementation_name != "pypy"
typing-extensions==4.10.0
tzdata==2024.1
win32-setctime==1.1.0; sys_platform == "win32"
